fix(concert): cancel request's 2 queries:change popolate function to aggregate function
