--- apps/2CD/concert/backend/src/resolvers/queries/requests/get-all-requists.ts
@@ -1,30 +1,0 @@
-import { QueryResolvers } from 'src/generated';
-import { RequestModel } from 'src/models';
-export const getAllRequists: QueryResolvers['getAllRequists'] = async () => {
-  const docs = await RequestModel.find()
-    .populate('booking')
-    .populate('user')
-    .populate({
-      path: 'booking',
-      populate: {
-        path: 'tickets.ticket',
-        model: 'Ticket',
-      },
-    })
-    .populate({
-      path: 'booking',
-      populate: { path: 'concert', model: 'Concert' },
-    });
-
-  return docs.map((doc) => ({
-    id: doc.id,
-    booking: doc.booking,
-    user: doc.user,
-    status: doc.status,
-    bank: doc.bank,
-    bankAccount: doc.bankAccount,
-    name: doc.name,
-    createdAt: doc.createdAt,
-    updatedAt: doc.updatedAt,
-  }));
-};
