+++ apps/2CD/concert/backend/src/resolvers/queries/requests/index.ts
@@ -1,2 +1,2 @@
 export * from './get-all-pending-request';
-export * from './get-all-requists';
+export * from './get-all-requests';
