+++ apps/2CD/concert/backend/src/schemas/index.ts
@@ -4,7 +4,7 @@ import { userDefs } from './user.schema';
 import { bookingDef } from './booking.schema';
 import { concertDef } from './concert.schema';
 import { otpDef } from './otp.schema';
-import { requestDef } from './requist.schema';
+import { requestDef } from './request.schema';
 import { scheduleDef } from './schedule.schema';
 import { ticketDef } from './ticket.schema';
 import { venueDef } from './venue.schema';
