+++ apps/2CD/concert/backend/src/schemas/request.schema.ts
@@ -20,7 +20,7 @@ export const requestDef = gql`
 
   type Query {
     getPendingRequests: [Request!]!
-    getAllRequists: [Request!]!
+    getAllRequests: [Request!]!
   }
   input updateReqInput {
     id: ID!
